Changed the way custom properties are exposed to prevent misuse/abuse.
No longer able to add custom device properties with keys that don't match the full property name (API simply doesn't allow it). Can still become out of sync if the property name or namespace changes after the property was added to a collection, but this is documented and should be rare.